+++ src/luxonis_train/models/modules/__init__.py
@@ -1,2 +1,2 @@
-from .common import *
+from .blocks import *
 from .activations import *
